DataOps/E2E-DataOps/requirement-dev.txt
# requirements-dev.txt
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pytest-cov==4.1.0
pre-commit==3.6.0
bandit==1.7.5